requirements.txt
xlrd>=0.7.8
django-redis-sessions
python-memcached
tilestache
modestmaps
django-celery<3.0.0
celery<=2.5.5
